account/page.tsxのリファクタリング
