+++ front/src/context/ConfigContext.tsx
@@ -15,7 +15,9 @@ type ConfigContextType = {
   setConfig: React.Dispatch<React.SetStateAction<ConfigItem[]>>;
 };
 
-const endpoint = `${process.env.NEXT_PUBLIC_API_ENDPOINT_URL}/config`;
+// const endpoint = `${process.env.NEXT_PUBLIC_API_ENDPOINT_URL}/config`;
+const endpoint = "http://localhost:3000/api/v1/config";
+
 const ConfigContext = createContext<ConfigContextType | null>(null);
 
 export const ConfigProvider = ({ children }: { children: ReactNode }) => {
